Struct engine::store::storage::cache::Cache

pub struct Cache<K, V> where
    K: Hash + Eq,
    V: Clone + Debug,  {
    table: HashMap<K, Value<V>>,
    scan_freq: Option<Duration>,
    created_at: SystemTime,
    last_scan_at: Option<SystemTime>,
}

The Cache struct used to store the data in an ordered format.

Fields
table: HashMap<K, Value<V>>scan_freq: Option<Duration>created_at: SystemTimelast_scan_at: Option<SystemTime>
Implementations

impl<K: Hash + Eq, V: Clone + Debug> Cache<K, V>

pub fn new() -> Self

creates a new empty Cache

Example

use engine::store::Cache;
use std::time::Duration;

let mut cache = Cache::new();

let key: Vec<u8> = b"key".to_vec();
let value: Vec<u8> = b"value".to_vec();

cache.insert(key.clone(), value.clone(), None);

assert_eq!(cache.get(&key), Some(&value))

pub fn create_with_scanner(scan_freq: Duration) -> Self

creates an empty Cache with a periodic scanner which identifies expired entries.

Example

use engine::store::Cache;
use std::time::Duration;

let scan_freq = Duration::from_secs(60);

let mut cache = Cache::create_with_scanner(scan_freq);

let key: &'static str = "key";
let value: &'static str = "value";

cache.insert(key, value, None);

assert_eq!(cache.get(&key), Some(&value));

pub fn get(&self, key: &K) -> Option<&V>

Gets the value associated with the specified key.

Example

use engine::store::Cache;
use std::time::Duration;

let mut cache = Cache::new();

let key: &'static str = "key";
let value: &'static str = "value";

cache.insert(key, value, None);

assert_eq!(cache.get(&key), Some(&value))

pub fn get_or_insert<F>(
&mut self,
key: K,
func: F,
lifetime: Option<Duration>
) -> &V where
F: Fn() -> V,

Gets the value associated with the specified key. If the key could not be found in the Cache, creates and inserts the value using a specified func function. # Example

use engine::store::Cache;
use std::time::Duration;

let mut cache = Cache::new();

let key = "key";
let value = "value";

assert_eq!(cache.get_or_insert(key, move || value, None), &value);
assert!(cache.contains_key(&key));

pub fn insert(
&mut self,
key: K,
value: V,
lifetime: Option<Duration>
) -> Option<V>

Insert a key-value pair into the cache. If key was not present, a None is returned, else the value is updated and the old value is returned.

Example

use engine::store::Cache;
use std::time::Duration;

let mut cache = Cache::new();

let key: &'static str = "key";
let value: &'static str = "value";

let insert = cache.insert(key, value, None);

assert_eq!(cache.get(&key), Some(&value));
assert!(insert.is_none());

pub fn remove(&mut self, key: &K) -> Option<V>

Removes a key from the cache. Returns the value from the key if the key existed in the cache.

Example

use engine::store::Cache;
use std::time::Duration;

let mut cache = Cache::new();

let key: &'static str = "key";
let value: &'static str = "value";

let insert = cache.insert(key, value, None);
assert_eq!(cache.remove(&key), Some(value));
assert!(!cache.contains_key(&key));

pub fn contains_key(&self, key: &K) -> bool

pub fn get_last_scanned_at(&self) -> Option<SystemTime>

pub fn get_scan_freq(&self) -> Option<Duration>

Get the cache’s scan frequency.

Example

use engine::store::Cache;
use std::time::Duration;

let scan_freq = Duration::from_secs(60);

let mut cache = Cache::create_with_scanner(scan_freq);

let key: &'static str = "key";
let value: &'static str = "value";

cache.insert(key, value, None);

assert_eq!(cache.get_scan_freq(), Some(scan_freq));

pub fn clear(&mut self)

Clear the stored cache and reset.

fn try_remove_expired_items(&mut self, now: SystemTime)

attempts to remove expired items based on the current system time provided.