Trait engine::Blender

pub trait Blender {
    // Required methods
    fn name(&self) -> &'static str;
    fn load(&mut self, marker: usize, img: &Image);
    fn clean(&mut self);
    fn blend(
        &mut self,
        mode: BlendMode,
        destination: &Rectangle,
        overlay: (&Image, Position, f64, Option<usize>),
        base: (&Image, Position, f64, Option<usize>)
    ) -> Image;
    fn blend_damaged_into(
        &mut self,
        mode: BlendMode,
        damage: &Rectangle,
        destination: (&mut Image, Position),
        overlay: (&Image, Position, f64),
        base: (&Image, Position, f64)
    );
    fn blend_damaged(
        &mut self,
        mode: BlendMode,
        base: (&mut Image, Position, f64),
        overlay: (&Image, Position, f64),
        damage: &Rectangle
    );

    // Provided method
    fn blend_all(
        &mut self,
        destination: &Rectangle,
        children: Vec<(BlendMode, &Image, Position, f64, bool, Option<usize>), Global>
    ) -> Image { ... }
}

A blender blends multiple images together (composition)

Required Methods

fn name(&self) -> &'static str

Name of the blender

fn load(&mut self, marker: usize, img: &Image)

Marks that the image with the given marker won’t change until the next call of the clean method. That is important for caching these images into texture if hardware acceleration is used.

fn clean(&mut self)

Cleans the cache and thus forgets all images that got loaded through the load method.

fn blend( &mut self, mode: BlendMode, destination: &Rectangle, overlay: (&Image, Position, f64, Option<usize>), base: (&Image, Position, f64, Option<usize>) ) -> Image

Blends overlay on top of base with the given mode, positions and alphas. Creates a new image that is defined by destination.

fn blend_damaged_into( &mut self, mode: BlendMode, damage: &Rectangle, destination: (&mut Image, Position), overlay: (&Image, Position, f64), base: (&Image, Position, f64) )

Blends overlay on top of base with the given mode, positions and alphas. Renders only the damaged area defined by destination and adjusts base accordingly.

fn blend_damaged( &mut self, mode: BlendMode, base: (&mut Image, Position, f64), overlay: (&Image, Position, f64), damage: &Rectangle )

Blends the active image on top of the base image by mutating the passed reference

Provided Methods

fn blend_all( &mut self, destination: &Rectangle, children: Vec<(BlendMode, &Image, Position, f64, bool, Option<usize>), Global> ) -> Image

Blends all given children based on their properties into one image that fits the destination