Trait engine::vault::crypto_box::BoxProvider

pub trait BoxProvider: Sized + Ord + PartialOrd {
    fn box_key_len() -> usize;

    fn box_overhead() -> usize;

    fn box_seal(key: &Key<Self>, ad: &[u8], data: &[u8]) -> Result<Vec<u8>>;

    fn box_open(key: &Key<Self>, ad: &[u8], data: &[u8]) -> Result<Vec<u8>>;

    fn random_buf(buf: &mut [u8]) -> Result<()>;

    fn random_vec(len: usize) -> Result<Vec<u8>> { ... }
}

A provider interface between the vault and a crypto box. See libsodium’s secretbox for an example.

Required methods

fn box_key_len() -> usize

defines the key length for the BoxProvider.

fn box_overhead() -> usize

defines the size of the Nonce combined with the Ad for the BoxProvider.

fn box_seal(key: &Key<Self>, ad: &[u8], data: &[u8]) -> Result<Vec<u8>>

seals some data into the crypto box using the Key and the associated data.

fn box_open(key: &Key<Self>, ad: &[u8], data: &[u8]) -> Result<Vec<u8>>

opens a crypto box to get data using the Key and the associated data.

fn random_buf(buf: &mut [u8]) -> Result<()>

fills a buffer [&mut [u8]] with secure random bytes.

Provided methods

fn random_vec(len: usize) -> Result<Vec<u8>>

creates a vector with secure random bytes based off of an inputted usize length.