Module engine::vault::types::utils

Structs

BlobId
A blob identifier

ChainId
A chain identifier. Used to identify a set of transactions in a version.

ClientId
Client Id type used to identify a client.

Id
A generic Id type used as the underlying type for the ClientId and VaultId types.

RecordHint
a record hint. Used as a hint to what this data is.

RecordId
A record identifier. Contains a ChainID which refers to the “chain” of transactions in the Version.

TransactionId
A transaction identifier

Val
a big endian encoded number

VaultId
Vault Id type used to identify a vault.